requirements.txt
alembic==1.12.0; python_version >= "3.7"
annotated-types==0.5.0; python_version >= "3.7"
anyio==3.7.1; python_version >= "3.7"
certifi==2023.7.22; python_version >= "3.7"
charset-normalizer==3.2.0; python_full_version >= "3.7.0" and python_version >= "3.7"
click==8.1.7; python_version >= "3.8"
colorama==0.4.6; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.8" and python_full_version >= "3.7.0"
fastapi-sqlalchemy==0.2.1; python_version >= "3.7"
fastapi==0.103.1; python_version >= "3.7"
greenlet==3.0.0; platform_machine == "aarch64" and python_version >= "3.7" or platform_machine == "ppc64le" and python_version >= "3.7" or platform_machine == "x86_64" and python_version >= "3.7" or platform_machine == "amd64" and python_version >= "3.7" or platform_machine == "AMD64" and python_version >= "3.7" or platform_machine == "win32" and python_version >= "3.7" or platform_machine == "WIN32" and python_version >= "3.7"
h11==0.14.0; python_version >= "3.8"
idna==3.4; python_version >= "3.7"
kaleido==0.2.1
mako==1.2.4; python_version >= "3.7"
markupsafe==2.1.3; python_version >= "3.7"
numpy==1.26.0; python_version >= "3.11" and python_version < "3.13"
packaging==23.1; python_version >= "3.7"
pandas==2.1.0; python_version >= "3.9"
pillow==10.0.1; python_version >= "3.8"
plotly==5.17.0; python_version >= "3.6"
psycopg2==2.9.9; python_version >= "3.7"
pydantic-core==2.6.3; python_version >= "3.7"
pydantic==2.3.0; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.9" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.9"
python-dotenv==1.0.0; python_version >= "3.8"
pytz==2023.3.post1; python_version >= "3.9"
requests==2.31.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.9" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.9"
sniffio==1.3.0; python_version >= "3.7"
sqlalchemy==2.0.22; python_version >= "3.7"
starlette==0.27.0; python_version >= "3.7"
tenacity==8.2.3; python_version >= "3.7"
typing-extensions==4.7.1; python_version >= "3.7"
tzdata==2023.3; python_version >= "3.9"
urllib3==2.0.4; python_version >= "3.7"
uvicorn==0.23.2; python_version >= "3.8"
